dca/pyproject.toml
[tool.poetry]
authors = ["Oskar Korczak <oskar.korczak@neontri.com>"]
description = ""
name = "dca"
version = "0.1.0"

[tool.poetry.dependencies]
PyYAML = "^6.0"
pydantic = "^1.10.2"
python = "^3.8"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
simplejson = "^3.17.6"
functions-framework = "^3.2.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.15.2"
pytest = "^5.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pylint.MASTER]
disable = ["C0114", "C0115", "C0116", "R0801", "W0511"] # missing module, class, function docstring, duplicate-code, fixme
extension-pkg-whitelist = "pydantic"
max-line-length = 240

[tool.isort]
line_length = 99
profile = "black"

[tool.black]
line-length = 99
